requirements.txt
colorgram.py==1.2.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.0
Flask-WTF==1.0.1
numpy==1.23.0
Pillow==9.1.1
Werkzeug==2.2.2
